# src/Foundry.WebApi/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Foundry.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/accounts/validate-token": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Validates a personal access token for the given provider",
        "operationId": "ValidateToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateTokenRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ValidateTokenRequestBody": {
        "required": [
          "token",
          "baseUrl",
          "providerType"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "providerType": {
            "type": "string"
          }
        }
      },
      "ValidateTokenResponse": {
        "required": [
          "kind",
          "accountName",
          "missingScopes",
          "detectedProvider"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "accountName": {
            "type": [
              "null",
              "string"
            ]
          },
          "missingScopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "detectedProvider": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}